# uint256: optimize umul, MulOverflow, MulMod, MulDivOverflow