The API of `umul` is changed to reduce an array copy.

## Test
```
go test ./...
```
```
ok  	github.com/holiman/uint256	1.381s
```
## Benchmark
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
                                 │     old     │                 new                 │
                                 │   sec/op    │   sec/op     vs base                │
Mul/single/uint256-16              4.306n ± 2%   4.248n ± 1%   -1.36% (p=0.007 n=10)
Mul/single/big-16                  37.11n ± 2%   36.92n ± 1%        ~ (p=0.072 n=10)
MulOverflow/single/uint256-16      14.64n ± 2%   10.07n ± 1%  -31.22% (p=0.000 n=10)
MulOverflow/single/big-16          37.04n ± 2%   37.02n ± 2%        ~ (p=0.698 n=10)
MulMod/small/uint256-16            20.92n ± 1%   17.79n ± 1%  -14.99% (p=0.000 n=10)
MulMod/mod64/uint256-16            37.74n ± 1%   36.41n ± 1%   -3.52% (p=0.000 n=10)
MulMod/mod128/uint256-16           61.82n ± 1%   57.38n ± 2%   -7.19% (p=0.000 n=10)
MulMod/mod192/uint256-16           77.23n ± 1%   72.75n ± 1%   -5.80% (p=0.000 n=10)
MulMod/mod256/uint256-16           88.37n ± 1%   86.69n ± 2%   -1.90% (p=0.000 n=10)
MulMod/mod256/uint256r-16          42.05n ± 1%   38.02n ± 1%   -9.57% (p=0.000 n=10)
MulMod/small/big-16                36.24n ± 1%   36.25n ± 1%        ~ (p=0.812 n=10)
MulMod/mod64/big-16                58.35n ± 1%   58.09n ± 1%        ~ (p=0.781 n=10)
MulMod/mod128/big-16               194.7n ± 0%   194.5n ± 1%        ~ (p=0.362 n=10)
MulMod/mod192/big-16               223.4n ± 1%   223.6n ± 1%        ~ (p=0.566 n=10)
MulMod/mod256/big-16               258.7n ± 1%   258.6n ± 0%        ~ (p=0.956 n=10)
MulDivOverflow/small/uint256-16    23.77n ± 2%   22.72n ± 1%   -4.44% (p=0.000 n=10)
MulDivOverflow/div64/uint256-16    27.85n ± 1%   27.46n ± 1%   -1.40% (p=0.001 n=10)
MulDivOverflow/div128/uint256-16   45.97n ± 1%   43.82n ± 1%   -4.69% (p=0.000 n=10)
MulDivOverflow/div192/uint256-16   59.26n ± 0%   56.96n ± 1%   -3.87% (p=0.000 n=10)
MulDivOverflow/div256/uint256-16   76.16n ± 1%   73.23n ± 1%   -3.85% (p=0.000 n=10)
MulDivOverflow/small/big-16        56.44n ± 1%   57.05n ± 1%   +1.07% (p=0.006 n=10)
MulDivOverflow/div64/big-16        63.14n ± 0%   63.33n ± 1%        ~ (p=0.492 n=10)
MulDivOverflow/div128/big-16       194.7n ± 1%   196.4n ± 1%   +0.85% (p=0.000 n=10)
MulDivOverflow/div192/big-16       227.0n ± 0%   230.2n ± 1%   +1.43% (p=0.000 n=10)
MulDivOverflow/div256/big-16       260.6n ± 0%   260.9n ± 0%        ~ (p=0.567 n=10)
geomean                            58.68n        56.38n        -3.92%
```