@holiman 2024-05-13
The numbers look really nice!
 
```
[user@work uint256]$ benchstat umul.1 umul.2
name                             old time/op    new time/op    delta
Mul/single/uint256-8               12.2ns ±35%    13.0ns ±42%     ~     (p=0.780 n=10+10)
MulOverflow/single/uint256-8       43.9ns ±37%    25.3ns ±42%  -42.45%  (p=0.000 n=10+10)
MulMod/small/uint256-8             54.9ns ±54%    36.3ns ±23%  -33.93%  (p=0.013 n=10+9)
MulMod/mod64/uint256-8              101ns ±53%     105ns ±31%     ~     (p=0.579 n=10+10)
MulMod/mod128/uint256-8             205ns ±43%     181ns ±39%     ~     (p=0.211 n=10+10)
MulMod/mod192/uint256-8             212ns ±53%     216ns ±42%     ~     (p=0.494 n=10+10)
MulMod/mod256/uint256-8             269ns ±41%     198ns ±37%  -26.29%  (p=0.042 n=10+10)
MulMod/mod256/uint256r-8            131ns ±14%     111ns ±36%     ~     (p=0.085 n=10+10)
MulDivOverflow/small/uint256-8     70.4ns ±44%    47.9ns ±29%     ~     (p=0.053 n=10+9)
MulDivOverflow/div64/uint256-8     93.1ns ±42%    84.8ns ±46%     ~     (p=0.436 n=10+10)
MulDivOverflow/div128/uint256-8     117ns ±23%     123ns ±54%     ~     (p=0.631 n=10+10)
MulDivOverflow/div192/uint256-8     177ns ±54%     182ns ±41%     ~     (p=0.739 n=10+10)
MulDivOverflow/div256/uint256-8     215ns ±66%     216ns ±59%     ~     (p=0.631 n=10+10)
```